wire up db
