+++ feature/discussion/server/server_test.go
@@ -3,7 +3,6 @@ package server_test
 import (
 	"bytes"
 	"context"
-	"database/sql"
 	"encoding/json"
 	"fmt"
 	"net"
@@ -12,13 +11,12 @@ import (
 	"testing"
 	"time"
 
-	"github.com/google/uuid"
-	_ "github.com/jackc/pgx/v5/stdlib"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"github.com/theRAGEhero/Democracy-Routes/feature/discussion/server/httpapi"
 	apimodel "github.com/theRAGEhero/Democracy-Routes/feature/discussion/server/httpapi/model"
+	"github.com/theRAGEhero/Democracy-Routes/feature/discussion/server/testhelper"
 	"github.com/theRAGEhero/Democracy-Routes/feature/discussion/server/userhandler"
 	usermodel "github.com/theRAGEhero/Democracy-Routes/feature/discussion/server/userhandler/model"
 )
@@ -29,7 +27,7 @@ func TestServer(t *testing.T) {
 	t.Run("new user", func(t *testing.T) {
 		t.Parallel()
 
-		userH, err := userhandler.New(tmpDBInstance(t))
+		userH, err := userhandler.New(testhelper.TmpDB(t))
 		require.NoError(t, err, "creating user handler")
 
 		// Given a new user Dima is added.
@@ -51,7 +49,7 @@ func TestServer(t *testing.T) {
 	t.Run("user authorization", func(t *testing.T) {
 		t.Parallel()
 
-		userH, err := userhandler.New(tmpDBInstance(t))
+		userH, err := userhandler.New(testhelper.TmpDB(t))
 		require.NoError(t, err, "creating user handler")
 		api := httpApi(t)
 
@@ -83,7 +81,7 @@ func TestServer(t *testing.T) {
 	t.Run("new meeting", func(t *testing.T) {
 		t.Parallel()
 
-		userH, err := userhandler.New(tmpDBInstance(t))
+		userH, err := userhandler.New(testhelper.TmpDB(t))
 		require.NoError(t, err, "creating user handler")
 		api := httpApi(t)
 
@@ -164,49 +162,4 @@ func randomPort(tb testing.TB) int {
 	require.NoError(tb, l.Close(), "closing listener")
 
 	return l.Addr().(*net.TCPAddr).Port
-}
-
-func tmpDBInstance(tb testing.TB) *sql.DB {
-	tb.Helper()
-
-	db := tmpDB(tb)
-	prepareDB(tb, db)
-
-	return db
-}
-
-func tmpDB(tb testing.TB) *sql.DB {
-	tb.Helper()
-
-	db, err := sql.Open("pgx", "postgres://postgres:testing@localhost:5432/postgres")
-	require.NoError(tb, err, "connecting to db")
-	require.NoError(tb, db.Ping(), "pinging db")
-
-	tdbn := "db" + strings.ReplaceAll(uuid.NewString(), "-", "")
-
-	_, err = db.Exec("CREATE DATABASE " + tdbn)
-	require.NoError(tb, err, "creating temporary db")
-
-	tdb, err := sql.Open("pgx", "postgres://postgres:testing@localhost:5432/"+tdbn)
-	require.NoError(tb, err, "connecting to temporary db")
-	require.NoError(tb, tdb.Ping(), "pinging temporary db")
-
-	tb.Cleanup(func() {
-		require.NoError(tb, tdb.Close(), "closing temporary db connection")
-		_, err = db.Exec("DROP DATABASE " + tdbn)
-		require.NoError(tb, err, "dropping temporary db")
-		require.NoError(tb, db.Close(), "closing db connection")
-	})
-
-	return tdb
-}
-
-func prepareDB(tb testing.TB, db *sql.DB) {
-	tb.Helper()
-
-	_, err := db.Exec(`CREATE TABLE users (
-			id uuid PRIMARY KEY,
-			name text NOT NULL
-		)`)
-	require.NoError(tb, err, "creating users table")
 }
